# docs/05-technical/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Travel & Business Trip Management API
  description: |
    API Contract cho hệ thống quản lý công tác doanh nghiệp.
    - **Base URL:** `http://localhost:5000/api/v1`
    - **Auth:** JWT Bearer (Access Token 15m) + Refresh Token (httpOnly cookie, 7d)
    - **Tiền tệ:** `integer` VND nguyên — ví dụ `1500000` = 1.500.000 VNĐ
    - **Tham chiếu:** `architecture.md`, `data-model.md`, `business-rules.md`
    - **FIX-08:** SQLite mutations reserve the database writer before reads. See `concurrency.md`.
      Business audit/notification rows commit atomically; SSE is emitted after commit.
      Contention exhaustion returns 409 CONCURRENT_MODIFICATION; stale state returns 409 INVALID_STATUS_TRANSITION or TRIP_IMMUTABLE.
  version: "1.0.0"
  contact:
    name: Nhóm 11 — MIS3032_1
  license:
    name: Academic Use Only
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development

# ─────────────────────────────────────────────────────────────
# SECURITY SCHEMES
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token — expires 15 phút.
        Lấy từ `POST /auth/login` hoặc `POST /auth/refresh`.
        Lưu in-memory (Zustand), KHÔNG localStorage.

  # ─────────────────────────────────────────────────────────────
  # SCHEMAS
  # ─────────────────────────────────────────────────────────────
  schemas:
...
    UserRole:
      type: string
      enum: [EMPLOYEE, MANAGER, TRAVEL_ADMIN, FINANCE, ADMIN]
    JobGrade:
      type: string
      enum: [STAFF, MANAGER_GRADE, DIRECTOR]
      description: |
        STAFF → hotel max 1.000.000 VNĐ/đêm (BR-TR-01)
        MANAGER_GRADE → 1.800.000 VNĐ/đêm
        DIRECTOR → 3.000.000 VNĐ/đêm
    DestinationType:
      type: string
      enum: [TIER1_CITY, OTHER]
      description: |
        TIER1_CITY (HN, HCM, ĐN) → per diem 400.000 VNĐ/ngày (BR-TR-02)
        OTHER → 300.000 VNĐ/ngày
    TripStatus:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - MANAGER_REVIEWING
        - PENDING_ADMIN_APPROVAL
        - APPROVED
        - ONGOING
        - EXPENSE_DRAFT
        - EXPENSE_SUBMITTED
        - EXPENSE_APPROVED
        - EXPENSE_REJECTED
        - MANAGER_REAPPROVE
        - CLOSED
        - REJECTED
...
    ApprovalLevel:
      type: string
      enum: [LEVEL_1, LEVEL_2, MANAGER_REAPPROVE]
    ApprovalAction:
      type: string
      enum: [APPROVED, REJECTED]
    ExpenseStatus:
      type: string
      enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, CLOSED]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 45
        totalPages:
          type: integer
          example: 3
    ErrorResponse:
      type: object
      required: [error, message, requestId]
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
          description: Machine-readable error code
        message:
          type: string
          example: "departureDate phải là ngày trong tương lai"
        details:
          type: object
          additionalProperties: true
          nullable: true
        requestId:
          type: string
          example: req_abc123

    # ── Policy Check ───────────────────────────────────────────
    PolicyViolation:
      type: object
      properties:
        code:
          type: string
          enum:
            - POLICY_VIOLATION_ACCOMMODATION_OVER_BUDGET
            - POLICY_VIOLATION_PER_DIEM_EXCEEDED
            - URGENT_TRIP_NOTICE
            - POLICY_VIOLATION_BUDGET_THRESHOLD
        detail:
          type: string
        severity:
          type: string
          enum: [WARNING, BLOCKER]
        rule:
          type: string
          example: BR-TR-01
        limit:
          type: integer
          nullable: true
        actual:
          type: integer
          nullable: true
    PolicyCheckResult:
      type: object
      properties:
        passed:
          type: boolean
        violations:
          type: array
          items:
            $ref: '#/components/schemas/PolicyViolation'
        violationCount:
          type: integer
          minimum: 0
        requiresLevel2Approval:
          type: boolean
        checkedAt:
          type: string
          format: date-time

    # ── Approval Record ────────────────────────────────────────
    ApprovalRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        approver:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            role:
              $ref: '#/components/schemas/UserRole'
        approvalLevel:
          $ref: '#/components/schemas/ApprovalLevel'
        action:
          $ref: '#/components/schemas/ApprovalAction'
        comment:
          type: string
          nullable: true
        budgetSnapshot:
          type: integer
          description: Snapshot ngân sách tại thời điểm duyệt (VND)
        hadViolationsSnapshot:
          type: boolean
        actedAt:
          type: string
          format: date-time

    # ── Trip ───────────────────────────────────────────────────
    TripSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            department:
              type: string
              nullable: true
            jobGrade:
              $ref: '#/components/schemas/JobGrade'
        origin:
          type: string
          example: Hà Nội
        destination:
          type: string
          example: TP. Hồ Chí Minh
        destinationType:
          $ref: '#/components/schemas/DestinationType'
        departureDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
        tripDays:
          type: integer
          description: Server-computed — GENERATED ALWAYS AS (returnDate - departureDate + 1)
          readOnly: true
        purpose:
          type: string
        estimatedBudget:
          type: integer
          description: VND nguyên
        status:
          $ref: '#/components/schemas/TripStatus'
        isUrgent:
          type: boolean
          readOnly: true
          description: Server-computed từ working days diff (BR-TR-03)
        requiresLevel2:
          type: boolean
          readOnly: true
          description: Server-set bởi PolicyCheckEngine (BR-TR-04)
        submittedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TripDetail:
      allOf:
        - $ref: '#/components/schemas/TripSummary'
        - type: object
          properties:
            urgencyReason:
              type: string
              nullable: true
            approvalReasons:
              type: array
              description: "Snapshot lý do duyệt 2 cấp tại lúc submit (D-16). Rỗng nếu 1 cấp."
              items:
                $ref: '#/components/schemas/ApprovalReason'
            level1Approval:
              $ref: '#/components/schemas/Level1Approval'
              nullable: true
            policyCheckResult:
              $ref: '#/components/schemas/PolicyCheckResult'
              nullable: true
            approvalRecords:
              type: array
              items:
                $ref: '#/components/schemas/ApprovalRecord'
            approvedAt:
              type: string
              format: date-time
              nullable: true
            closedAt:
              type: string
              format: date-time
              nullable: true
    TripCreateRequest:
      type: object
      required:
        - origin
        - destination
        - departureDate
        - returnDate
        - purpose
        - estimatedBudget
      description: "D-16: destinationType và hotelCostPerNight/perDiemBudget không còn nhận từ client. Server tự tính destinationType từ destination."
      properties:
        origin:
          type: string
          minLength: 1
          maxLength: 200
          example: Hà Nội
        destination:
          type: string
          minLength: 1
          maxLength: 200
          example: Đà Nẵng
        departureDate:
          type: string
          format: date
          description: Phải >= ngày hôm nay
          example: "2026-10-01"
        returnDate:
          type: string
          format: date
          description: Phải >= departureDate
          example: "2026-10-03"
        purpose:
          type: string
          minLength: 10
          maxLength: 1000
        estimatedBudget:
          type: integer
          minimum: 1
          description: "Ngân sách dự kiến tổng chuyến đi (VND nguyên). Đây là Combined_Actual cho BR-TR-08."
        urgencyReason:
          type: string
          minLength: 10
          maxLength: 500
          description: "Bắt buộc nếu workingDays < 3 (BR-TR-03)"
          nullable: true
    TripUpdateRequest:
      type: object
      description: Tất cả fields optional. Server strip status/isUrgent/requiresLevel2/tripDays.
      properties:
        origin:
          type: string
          minLength: 1
          maxLength: 200
        destination:
          type: string
          minLength: 1
          maxLength: 200
        destinationType:
          $ref: '#/components/schemas/DestinationType'
        departureDate:
          type: string
          format: date
        returnDate:
          type: string
          format: date
        purpose:
          type: string
          minLength: 10
          maxLength: 1000
        estimatedBudget:
          type: integer
          minimum: 1
        hotelCostPerNight:
          type: integer
          minimum: 0
          nullable: true
        hotelNights:
          type: integer
          minimum: 0
          nullable: true
        perDiemBudget:
          type: integer
          minimum: 0
          nullable: true
        transportBudget:
          type: integer
          minimum: 0
          nullable: true
        otherBudget:
          type: integer
          minimum: 0
          nullable: true

    # ── Itinerary ──────────────────────────────────────────────
...
  responses:
    MutationConflict:
      description: Stale state, CLOSED trip, exhausted concurrency retry or changed payload for an existing Idempotency-Key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            contention:
              value: { error: CONCURRENT_MODIFICATION, message: 'Dữ liệu đang được cập nhật. Vui lòng thử lại.', details: {}, requestId: req_abc123 }
            idempotency:
              value: { error: IDEMPOTENCY_CONFLICT, message: 'Idempotency-Key đã được dùng với nội dung khác.', details: {}, requestId: req_abc123 }
    Unauthorized:
      description: Token thiếu hoặc không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Token không hợp lệ hoặc đã hết hạn
            requestId: req_abc123
    Forbidden:
      description: Không đủ quyền
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: "Role 'EMPLOYEE' không có quyền thực hiện action này"
            requestId: req_abc123
    NotFound:
      description: Resource không tồn tại
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Trip không tồn tại
            requestId: req_abc123
    ValidationError:
      description: Input không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: departureDate phải là ngày trong tương lai
            details:
              field: departureDate
              received: "2025-01-01"
            requestId: req_abc123
...
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Same actor/operation/resource/key and payload replay the committed result. Different payload returns 409. Use a new key for a new intent; no content-based item deduplication. Retain the key on retry.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    tripId:
      name: tripId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

# ─────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────
paths:
...
  /trips:
    get:
      tags: [Trips]
      summary: Danh sách Trip Requests
      operationId: listTrips
      security:
        - BearerAuth: []
      description: |
        Kết quả được filter server-side theo role:
        - EMPLOYEE: chỉ trip của mình
        - MANAGER: trip của subordinates
        - TRAVEL_ADMIN / FINANCE / ADMIN: tất cả
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TripStatus'
        - name: employeeId
          in: query
          description: Chỉ MANAGER+ mới dùng được
          schema:
            type: string
            format: uuid
        - name: departureFrom
          in: query
          schema:
            type: string
            format: date
        - name: departureTo
          in: query
          schema:
            type: string
            format: date
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, departureDate, estimatedBudget]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Danh sách trips
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TripSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Trips]
      summary: Tạo Trip Request mới
      operationId: createTrip
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      security:
        - BearerAuth: []
      description: |
        **Roles:** EMPLOYEE only.
        Server-computed (bị strip nếu client gửi): `tripDays`, `isUrgent`, `requiresLevel2`, `status`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreateRequest'
      responses:
        '201':
          description: Trip được tạo với status DRAFT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}:
    parameters:
      - $ref: '#/components/parameters/tripId'

    get:
      tags: [Trips]
      summary: Chi tiết Trip Request
      operationId: getTripById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Trip detail với policyCheckResult và approvalRecords
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Trips]
      summary: Cập nhật Trip Request
      operationId: updateTrip
      security:
        - BearerAuth: []
      description: |
        **Roles:** EMPLOYEE owner.
        **Điều kiện:** status === DRAFT.
        Server strip: `status`, `isUrgent`, `requiresLevel2`, `tripDays`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdateRequest'
      responses:
        '200':
          description: Trip đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'

    delete:
      tags: [Trips]
      summary: Xóa Trip Request
      operationId: deleteTrip
      security:
        - BearerAuth: []
      description: Chỉ xóa được khi status === DRAFT.
      responses:
        '204':
          description: Đã xóa
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}/submit:
    post:
      tags: [Trip Actions]
      summary: Nộp Trip Request
      operationId: submitTrip
      security:
        - BearerAuth: []
      description: |
        **Roles:** EMPLOYEE owner. **Điều kiện:** status === DRAFT.
        Server chạy PolicyCheckEngine (BR-TR-01, 02, 03, 04), set isUrgent + requiresLevel2.
      parameters:
        - $ref: '#/components/parameters/tripId'
      responses:
        '200':
          description: Trip đã submit, kèm kết quả policy check
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/TripStatus'
                  isUrgent:
                    type: boolean
                  requiresLevel2:
                    type: boolean
                  submittedAt:
                    type: string
                    format: date-time
                  policyCheckResult:
                    $ref: '#/components/schemas/PolicyCheckResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}/approve:
    post:
      tags: [Trip Actions]
      summary: Phê duyệt Trip Request
      operationId: approveTrip
      security:
        - BearerAuth: []
      description: |
        **Roles:**
        - MANAGER → Level 1 (khi status = SUBMITTED)
        - TRAVEL_ADMIN → Level 2 (khi status = PENDING_ADMIN_APPROVAL)

        **BR-TR-04:** Nếu budget > 20M hoặc có violation → PENDING_ADMIN_APPROVAL (cần L2).
      parameters:
        - $ref: '#/components/parameters/tripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  maxLength: 1000
                  example: Phê duyệt chuyến công tác
      responses:
        '200':
          description: Trip đã được duyệt
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/TripStatus'
                  approvalRecord:
                    $ref: '#/components/schemas/ApprovalRecord'
                  approvedAt:
                    type: string
                    format: date-time
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}/reject:
    post:
      tags: [Trip Actions]
      summary: Từ chối Trip Request
      operationId: rejectTrip
      security:
        - BearerAuth: []
      description: |
        **Roles:** MANAGER (L1), TRAVEL_ADMIN (L2).
        `comment` là bắt buộc khi reject.
      parameters:
        - $ref: '#/components/parameters/tripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment:
                  type: string
                  minLength: 10
                  maxLength: 1000
                  example: Ngân sách vượt kế hoạch quý, đề nghị điều chỉnh
      responses:
        '200':
          description: Trip đã bị từ chối
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/TripStatus'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}/close:
    post:
      tags: [Trip Actions]
      summary: Đóng hồ sơ Trip (Finance)
      operationId: closeTrip
      security:
        - BearerAuth: []
      description: |
        **Roles:** FINANCE.
        **Điều kiện:** `expense.status === APPROVED` (không phải trip.status).
        Nếu `managerReapprovalRequired = true` thì `managerReapproved` phải là `true` trước.
        **BR-TR-06:** Trip sau khi CLOSED là immutable hoàn toàn.

        **⚠️ UX vs API (L-05 fix):** Frontend có thể hiển thị nút "Approve & Close" gộp,
        nhưng phải gọi **tuần tự 2 request**:
        1. `POST /trips/:id/expense/approve` — chờ `expense.status = APPROVED`
        2. `POST /trips/:id/close` — chỉ gọi sau khi bước 1 thành công

        Cả `trip.status` và `expense.status` đều chuyển sang `CLOSED` trong 1 transaction.
      parameters:
        - $ref: '#/components/parameters/tripId'
      responses:
        '200':
          description: Trip và Expense đã đóng — cả hai immutable (BR-TR-06)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/TripStatus'
                  closedAt:
                    type: string
                    format: date-time
                  expense:
                    type: object
                    description: Expense cũng đồng thời chuyển sang CLOSED
                    properties:
                      id:
                        type: string
                        format: uuid
                      status:
                        $ref: '#/components/schemas/ExpenseStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'

        '422':
          description: Variance > 10% chưa được Manager re-approve (BR-TR-05)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: EXPENSE_VARIANCE_EXCEEDED
                message: "Variance 15.3% > 10%. Manager chưa phê duyệt phần vượt chi (BR-TR-05)."
                requestId: req_abc123

  # ── Itinerary ─────────────────────────────────────────────────
...